+++ lib/QECCore/src/QECCore.jl
@@ -15,7 +15,7 @@ export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanc
 export Perfect5, Cleve8, Gottesman
 
 # CSS Codes
-export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, DelfosseReichardt, QuantumTannerGraphProduct, CyclicQuantumTannerGraphProduct
+export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, DelfosseReichardt, DelfosseReichardtRepCode, QuantumTannerGraphProduct, CyclicQuantumTannerGraphProduct
 
 # Classical Codes
 export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
@@ -51,6 +51,7 @@ include("codes/quantum/quantumreedmuller.jl")
 
 # Delfosse-Reichardt Codes
 include("codes/quantum/delfosse_reichardt_code.jl")
+include("codes/quantum/delfosse_reichardt_repcode.jl")
 
 function __init__()
     if isdefined(Base.Experimental, :register_error_hint)