+++ lib/QECCore/test/test_codes.jl
@@ -59,6 +59,10 @@
         include("codes/hamming.jl")
     end
 
+    @testset "Delfosse-Reichardt Repetition Code" begin
+        include("codes/delfosse_reichardt_repcode.jl")
+    end
+
     @testset "Delfosse-Reichardt Code" begin
         include("codes/delfosse_reichardt_code.jl")
     end