+++ test/test_ecc_base.jl
@@ -251,7 +251,8 @@ const code_instance_args = Dict(
     :Triangular488 => [3, 5, 7, 9, 11],
     :Triangular666 => [3, 5, 7, 9, 11],
     :DelfosseReichardt => [(2,1,3), (2,2,4), (4,3,5), (4,3,6)],
-    :QuantumTannerGraphProduct => [(H1, H2),(H2, H2), (H1, H1), (H2, H1)],
+    :DelfosseReichardtRepCode => [4, 6, 8, 10],
+    :QuantumTannerGraphProduct => [(H1, H2),(H2, H2), (H1, H1), (H2, H1)]
 )
 
 function all_testablable_code_instances(;maxn=nothing)